refactor: Header에 있던 login, signup 관련 상태 및 로직을 LoginContext로 이동
- 그 외 LoginModalContext 사용하던 컴포넌트가 LoginContext 사용하도록 수정
