+++ src/contexts/LoginContext.jsx
@@ -7,8 +7,28 @@ export const useLoginContext = () => useContext(LoginContext);
 
 const LoginContextProvider = ({ children }) => {
   const [loggedIn, setLoggedIn] = useState(false);
+  const [clickedSignup, setClickedSignup] = useState(false);
+  const [clickedSignin, setClickedSignin] = useState(false);
   const [userInfo, setUserInfo] = useState({});
   const { id, nickname, profileImage } = userInfo;
+
+  const handleSigninModal = type => {
+    switch (type) {
+      case 'OPEN':
+        setClickedSignin(true);
+        break;
+      case 'CLOSE':
+        setClickedSignin(false);
+        break;
+      default:
+        setClickedSignin(!clickedSignin);
+        break;
+    }
+  };
+
+  const toggleSignupModal = () => {
+    setClickedSignup(!clickedSignup);
+  };
 
   useEffect(() => {
     (async () => {
@@ -29,7 +49,11 @@ const LoginContextProvider = ({ children }) => {
         setLoggedIn,
         id,
         nickname,
-        profileImage
+        profileImage,
+        clickedSignup,
+        clickedSignin,
+        handleSigninModal,
+        toggleSignupModal
       }}
     >
       {children}
